S3b: the C++ fire LOGISTIC in integer Q16.16 + sqrt_q16 (the core; FINAL field migration)
The second sub-step of S3 (fire -> integer) and the LAST per-cell-float removal of
the fixed-point arc. Converts FireSimulation::step's signed-logistic update to
integer Q16.16 end-to-end and adds the arc's FIRST per-cell transcendental
(sqrt_q16). Branch s3-fire-fixedpoint, continuing from S3a 28d134b; determinism-
critical (the core fire math + the extinguish/burn-through discrete outputs).

The new primitive (fixed_point.h)
- sqrt_q16(int64 x_q32) -> q16: a FIXED-32-iteration integer floor-isqrt of the
  int64 Q.32 radicand (wx^2+wy^2). isqrt of a Q.32 value yields Q16.16 directly
  (sqrt(2^32)=2^16, no rescale). Exact floor, branch-identical, no LUT/poly/libm,
  identical CPU<->__device__. Self-guards a radicand >= 2^62 (root >= 2^31) by
  clamping to INT32_MAX (dead on the real call site, |wind| is O(1)). Erik ratified
  FLOOR-isqrt (the 1-LSB bias is on the gentle wind-fan/strip tuning terms).
- narrow_round / narrow_round_signed: the shared round-to-nearest DEPOSIT narrows
  (one definition of the deterministic deposit rounding for fire + atmosphere).
- Unit gate tests/test_s3b_sqrt.py (+ _s3b_sqrt_check.cpp): r^2 <= x < (r+1)^2 over
  a sampled int64 domain incl. 0 + near-2^62; the scale-fold; determinism; the
  clamp guard; + the narrow_round helpers vs a double reference.

The integer logistic (fire_simulation.cpp)
- I<=0 skip; max_fire<0.001 early-exit -> integer max + quantize compare. T is the
  raw temperature[i] (temp_scale==FP_ONE -> the divide is the identity; a load-time
  reciprocal fallback handles a non-default temp_scale, so the identity is NOT
  guarded by a Release-stripped assert). F = clamp01(recip_mul(wall_hp, recip_fuel_ref)).
  P = mean_round(sum_atm, count) over open (~wall & ~vacuum) neighbours.
  W = sqrt_q16(mul_wide(wx,wx)+mul_wide(wy,wy)). hot/o2(smoothstep, pinned Hermite
  order)/avail. The 6-factor grow + 2-term die are a PINNED left-fold of mul_q16
  (deterministic by construction), narrowed at I_next = clamp01(I + mul_q16(dt,
  grow-die)); I_min snap-extinguish. Plume/smoke/wall-burn deposits ROUND-TO-NEAREST
  (unbiased sources). All divides are by config constants -> load-time reciprocals.

The C++ P gate adopts mean_round + the Python-twin negative-branch fix (review carry-forwards)
- The C++ fire P gate now uses int64 sum + mean_round (closing the S3a exact-0.60-tie
  gap so fire + the Python ignition gate fire IDENTICALLY).
- mean_round truncates toward 0; the S3a Python twin used // (floor toward -inf) ->
  they diverged on a NEGATIVE neighbour sum (atmosphere can dip transiently negative).
  combat.apply_temperature_ignition now emulates trunc-toward-zero on the negative
  branch. tests/test_fire_o2_invariant.py adds the EXACT-agreement gate (ZERO
  disagreements incl. negative-atmosphere configs).

wall_hp -> int32 Q16.16 (CPU)
- new src/simulation/wall_fixed.py (int16 Q8.8 GPU width recorded in the format tag).
  gamemap allocation/cache/patch quantize; physics.apply_blast_damage quantizes the
  decrement; every reader dequantizes (test_materials, test_door_explosive sums in
  int64). bindings + step_tail wall_hp arg -> int32.

The bridge reduction (physics_engine.cpp step_tail)
- fire reads atmosphere/wind/fire/wall_hp as int32 directly + writes the int32
  atmosphere plume in place -> the S3a fire-field bridge + the S2c atm/wind float
  bridges that fed the fire are GONE. The LAST float bridge is the temperature pass's
  POST-plume atmosphere read (atm_f_) -> S3c. (wind_x_f_/wind_y_f_/fire_f_ are now
  dead, kept declared for S3c's clean deletion.)

Gates (all green)
- P1 self-match tol=0.0 (fire + wall_hp int32, bit-identical run-to-run).
- Discrete-output determinism (test_s3b_fire_determinism.py): the extinguish-flip
  tick + the burn-through destroyed-tile list bit-identical run-twice across an
  ignite->firestorm->starve->extinguish trajectory; a shockwave-fanned firestorm
  (|wind|~30, violent wave) never overflows the Q16.16 [0,1] range; cross-config
  self-match (varied dt + [physics.fire] params).
- Feel-A/B (_s3b_firestorm_feel.{py,pkl,png}, committed artifact): float golden vs
  integer build -> fire max-cell |delta| 1.37e-4, L2 2.3e-3, mean-signed-error
  -9.8e-7; the atmosphere PLUME mean-signed-error -8.6e-9 (unbiased deposit);
  ignition + extinguish ticks match exactly.
- no-float ratchet: fire_simulation.cpp float 31->6 (logistic integer), double 2->19
  (load-time constant precompute); physics_engine.cpp float 66->65, double 32->30.
- Full suite 365 green; both test_main_smoke.py --auto (default + unhcr_vessel_2) exit 0.

S3c (next): delete the wind/fire scratch buffers + the temperature-pass atmosphere
float bridge (temperature_solver atmosphere arg -> int32), then fire+temperature join
the hard 0/0/0 no-float gate.

+++ cpp/src/bindings.cpp
@@ -202,13 +202,13 @@ PYBIND11_MODULE(breach_physics, m) {
         .def(py::init<>())
         .def_readwrite("params", &FireSimulation::params)
         .def("step", [](const FireSimulation& self,
-                        py::array_t<float> fire,
-                        py::array_t<float> atmosphere,
+                        py::array_t<int32_t> fire,         // S3b: Q16.16 int32
+                        py::array_t<int32_t> atmosphere,   // S2c: Q16.16 int32
                         py::array_t<int32_t> smoke,        // S2b: Q16.16 int32
-                        py::array_t<float> wall_hp,
+                        py::array_t<int32_t> wall_hp,      // S3b: Q16.16 int32
                         py::array_t<int32_t> temperature,
-                        py::array_t<float> wind_x,
-                        py::array_t<float> wind_y,
+                        py::array_t<int32_t> wind_x,       // S2c: Q16.16 int32
+                        py::array_t<int32_t> wind_y,       // S2c: Q16.16 int32
                         py::array_t<bool>  is_wall,
                         py::array_t<bool>  is_vacuum,
                         py::array_t<bool>  flammable,
@@ -553,12 +553,12 @@ PYBIND11_MODULE(breach_physics, m) {
                              py::array_t<int32_t> wave_p,        // S2a: Q16.16 int32
                              py::array_t<bool>  solid,
                              // fire group
-                             py::array_t<int32_t> fire,          // S3a: Q16.16 int32 (fire field bridge inside)
-                             py::array_t<int32_t> atmosphere,   // S2c: Q16.16 int32 (fire bridge inside)
+                             py::array_t<int32_t> fire,          // S3b: Q16.16 int32 (integer logistic)
+                             py::array_t<int32_t> atmosphere,   // S2c: Q16.16 int32 (fire reads + plume-writes)
                              py::array_t<int32_t> smoke,         // S2b: Q16.16 int32
-                             py::array_t<float> wall_hp,
+                             py::array_t<int32_t> wall_hp,       // S3b: Q16.16 int32
                              py::array_t<int32_t> temperature,
-                             py::array_t<int32_t> wind_x,        // S2c: Q16.16 int32 (fire bridge inside)
+                             py::array_t<int32_t> wind_x,        // S2c: Q16.16 int32
                              py::array_t<int32_t> wind_y,        // S2c: Q16.16 int32
                              py::array_t<bool>  is_vacuum,
                              py::array_t<bool>  flammable,
